wip: Handle reading taxids from NCBI taxdump database

+++ migrations/000003_initial_bgc_entry.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS mibig.entries (
     entry_id text PRIMARY KEY,
     minimal bool NOT NULL,
-    tax_id int NOT NULL,
+    tax_id bigint NOT NULL,
     organism_name text NOT NULL,
     biosyn_class text[] NOT NULL,
     legacy_comment text
